Add cascading menus to canned responses
