+++ plugin-flex-ts-template-v2/src/feature-library/canned-responses/flex-hooks/paste-elements/index.ts
@@ -3,7 +3,6 @@ import { PasteCustomCSS } from '@twilio-paste/customization';
 export const pasteElementHook = {
   CANNED_RESPONSES_MENU: {
     maxHeight: '70vh',
-    height: '100%',
     overflowY: 'scroll',
   },
   CANNED_RESPONSES_MENU_BUTTON: {
